tests/decode-split-character.html

<!DOCTYPE html>
<script src="resources/text-encode-transform.js"></script>
<script src="resources/transform-stream.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/readable-stream-from-array.js"></script>
<script src="resources/readable-stream-to-array.js"></script>
<script>
'use strict';

const inputBytes = [73, 32, 240, 159, 146, 153, 32, 115, 116, 114, 101,
                    97, 109, 115];
const expectedOutput = ['I ', '\u{1F499} streams'];

for (const splitPoint of [2, 3, 4, 5]) {
  promise_test(() => {
    const input = readableStreamFromArray(
        [new Uint8Array(inputBytes.slice(0, splitPoint)),
         new Uint8Array(inputBytes.slice(splitPoint))]);
    const output = input.pipeThrough(new TextDecoder());
    return readableStreamToArray(output).
        then(array => {
          assert_array_equals(
              array, expectedOutput,'the split code point should be in the ' +
                  'second chunk of the output');
        });
  }, 'a code point split between chunks should not be emitted until all ' +
      'bytes are available; split point = ' + splitPoint);
}

promise_test(() => {
  const splitPoint = 6;
  const input = readableStreamFromArray(
      [new Uint8Array(inputBytes.slice(0, splitPoint)),
       new Uint8Array(inputBytes.slice(splitPoint))]);
  const output = input.pipeThrough(new TextDecoder());
  return readableStreamToArray(output).
      then(array => {
        assert_array_equals(array, ['I \u{1F499}', ' streams'],
                            'the multibyte character should be in the ' +
                                'first chunk of the output');
      });
}, 'a code point should be emitted as soon as all bytes are available');
</script>
